% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Shiny_qraLm.R
\name{run_Shiny_qraLm}
\alias{run_Shiny_qraLm}
\title{Launch the Shiny_qraLm application}
\usage{
run_Shiny_qraLm(...)
}
\arguments{
\item{...}{Additional parameters passed to \code{shiny::runApp()}.}
}
\description{
Launch the Shiny_qraLm application
}
